requirements.txt
fastapi
uvicorn[standard]
pydantic
openai-whisper
pytz
psycopg2-binary
python-multipart
httpx
pytest
sqlalchemy